Products Module
Design and coding of the listing page of the products module has been completed. In this context, a database connection was established and the page was created with dynamic data. DataTables integration builded for the table structure on the page. Made DataTables Turkish with language json file. The design and coding of the Add New Products page has been completed. Form Validation library was included in the project. A helper named tools_helper was defined to identify the extra functions needed in the application. In this helper, the function called ConvertToSEO was defined and the SLUG creation process was completed. The products module will be continued with the integration of editing, deleting, dynamic sorting, sweet alert and notifications.